[#19] feat: udp 데이터 통신 방식 변경 (한 개의 포트로 모든 입력 수용)

+++ logic-server/logic-server/main.cpp
@@ -18,8 +18,10 @@ int main()
     const std::size_t rpcCtxThreadCount = ctxThreadCount / 5; // 20% of total threads for RPC
     const std::size_t workCtxThreadCount = ctxThreadCount - rpcCtxThreadCount; // Remaining threads for work context
 
-    auto workThreadContext = ContextManager::Create("work", workCtxThreadCount * 0.3f, workCtxThreadCount * 0.7f); // work thread for normal network callback
-    auto rpcThreadContext = ContextManager::Create("rpc", rpcCtxThreadCount * 0.3f, rpcCtxThreadCount * 0.7f); // rpc thread for udp network callback
+    auto workThreadContext = ContextManager::Create("work", 
+        static_cast<std::size_t>(workCtxThreadCount * 0.3f), static_cast<std::size_t>(workCtxThreadCount * 0.7f)); // work thread for normal network callback
+    auto rpcThreadContext = ContextManager::Create("rpc",
+        static_cast<std::size_t>(rpcCtxThreadCount * 0.3f), static_cast<std::size_t>(rpcCtxThreadCount * 0.7f)); // rpc thread for udp network callback
 
     auto internalConnector = std::make_shared<InternalConnector>();
     if (!internalConnector->GetAccessTokenFromInternal())
